move field types to sub dir
