+++ packages/types/src/entities/fields/Name.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023 The Ontario Institute for Cancer Research. All rights reserved
+ * Copyright (c) 2024 The Ontario Institute for Cancer Research. All rights reserved
  *
  * This program and the accompanying materials are made available under the terms of
  * the GNU Affero General Public License v3.0. You should have received a copy of the
@@ -19,7 +19,7 @@
 
 import { z } from 'zod';
 
-const HISTORY_OF_CANCER = ['YES', 'NO', 'UNKNOWN'] as const;
+import { name } from '../../common/Regex.js';
 
-export const HistoryOfCancer = z.enum(HISTORY_OF_CANCER);
-export type HistoryOfCancer = z.infer<typeof HistoryOfCancer>;
+export const Name = z.string().trim().regex(name);
+export type Name = z.infer<typeof Name>;
