+++ packages/types/src/entities/fields/PostalCode.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023 The Ontario Institute for Cancer Research. All rights reserved
+ * Copyright (c) 2024 The Ontario Institute for Cancer Research. All rights reserved
  *
  * This program and the accompanying materials are made available under the terms of
  * the GNU Affero General Public License v3.0. You should have received a copy of the
@@ -19,7 +19,10 @@
 
 import { z } from 'zod';
 
-import { name } from './Regex.js';
+import { postalCode } from '../../common/Regex.js';
 
-export const Name = z.string().trim().regex(name);
-export type Name = z.infer<typeof Name>;
+export const PostalCode = z
+	.string()
+	.regex(postalCode)
+	.transform((data) => data.toUpperCase());
+export type PostalCode = z.infer<typeof PostalCode>;
