+++ packages/types/src/entities/fields/index.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023 The Ontario Institute for Cancer Research. All rights reserved
+ * Copyright (c) 2024 The Ontario Institute for Cancer Research. All rights reserved
  *
  * This program and the accompanying materials are made available under the terms of
  * the GNU Affero General Public License v3.0. You should have received a copy of the
@@ -17,9 +17,19 @@
  * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import { ZodError } from 'zod';
-
-// TODO: remove and use custom Error schemas
-// this is just here for temporary error handling in consent-api to return a 400 status for invalid requests
-// when we implement custom error schemas this will not need to be exported
-export { ZodError };
+export * from './Ancestry.js';
+export * from './BirthSex.js';
+export * from './ConsentCategory.js';
+export * from './ConsentGroup.js';
+export * from './Gender.js';
+export * from './GeneticsClinic.js';
+export * from './HistoryOfCancer.js';
+export * from './LifecycleState.js';
+export * from './MolecularLab.js';
+export * from './Name.js';
+export * from './NanoId.js';
+export * from './OhipNumber.js';
+export * from './PhoneNumber.js';
+export * from './PostalCode.js';
+export * from './Province.js';
+export * from './UserRole.js';
